SKS TA relies on the 32bit version of the cryptoki IDs and structure
The PKCS#11 2.40 does not specify whether the 'unsigned long int' type
is 32bit or 64bit. The SKS TA relies on the cryptoki IDs and structures
as defined by the PKCS#11. SKS client application may be 32bit or 64bit,
SKS TA itself can be 32bit or 64bit. To prevent from inconsistencies,
the SKS TA always rely on 32bit 'int' configuration. SKS TA can send
cyrptoki formatted structure but client library will have to convert
between the 32bit configuration of the TA and the 32bit or 64bit
configuration of the client application.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>