sks abi: vendor IDs for undefined ID and secure storage ID.
SKS_UNDEFINED_ID and CKA_SKS_SECSTOR_DATA is a vendor ID defined by
the Secure Key Services.

SKS_UNDEFINED_ID specify an undefined value is a object. This specific
value is meaningful only for specific object attributes as the object
class, or an object type-in-class when such information is needed. This
value is reserved in the ABI and used in SKS attribute parsing sequences.

Attribute CKA_SKS_SECSTOR_DATA is used to store a reference to the object
data when the object data is stored in outside the object attributes as
a standlone object in the SKS persistent secure storage.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>